paushing/resuming on bot disconnect/connect works now but I used an evil method. Dunno if I can get it to look any better though
