+++ src/server/api/api.ts
@@ -22,6 +22,7 @@ import {
     gameManager,
     gamePaused,
     pauseGame,
+    setAllRobotsToDefaultPositions,
     setGameManager,
     socketManager,
     unpauseGame,
@@ -36,7 +37,7 @@ import { Side } from "../../common/game-types";
 import { USE_VIRTUAL_ROBOTS, START_ROBOTS_AT_DEFAULT } from "../utils/env";
 import { SaveManager } from "./save-manager";
 
-import { VirtualBotTunnel, VirtualRobot } from "../simulator";
+import { VirtualBotTunnel} from "../simulator";
 import { Position } from "../robot/position";
 import { DEGREE } from "../../common/units";
 import { PacketType } from "../utils/tcp-packet";
@@ -85,26 +86,6 @@ async function setupDefaultRobotPositions(
             await executor.execute(command);
         } else {
             setAllRobotsToDefaultPositions();
-        }
-    }
-}
-
-function setAllRobotsToDefaultPositions(
-    defaultPositionsMap?: Map<string, GridIndices>,
-): void {
-    if (defaultPositionsMap) {
-        for (const [robotId, indices] of defaultPositionsMap.entries()) {
-            const robot = robotManager.getRobot(robotId);
-            robot.position = Position.fromGridIndices(indices);
-            if (robot instanceof VirtualRobot)
-                robot.updateTunnelPosition(robot.position);
-        }
-    } else {
-        for (const robot of robotManager.idsToRobots.values()) {
-            robot.position = Position.fromGridIndices(robot.defaultIndices);
-            if (robot instanceof VirtualRobot)
-                robot.updateTunnelPosition(robot.position);
-            robotManager.updateRobot(robot.id, robot.defaultIndices);
         }
     }
 }
@@ -563,27 +544,6 @@ apiRouter.get("/get-puzzles", (_, res) => {
     return res.send(out);
 });
 
-export function doRollBack() {
-    const ids = clientManager.getIds();
-    if (ids) {
-        doRollBack();
-
-        const oldSave = SaveManager.loadGame(ids[0]);
-        gameManager?.chess.loadFen(oldSave!.oldPos);
-        setAllRobotsToDefaultPositions(
-            new Map(
-                oldSave!.oldRobotPos?.map<[string, GridIndices]>((obj) => [
-                    obj[1],
-                    new GridIndices(
-                        parseInt(obj[0].split(", ")[0]),
-                        parseInt(obj[0].split(", ")[1]),
-                    ),
-                ]),
-            ),
-        );
-        socketManager.sendToAll(new SetChessMessage(oldSave!.oldPos));
-    }
-}
 
 /**
  * Pause the game
